requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
cryptography==43.0.3
dataclasses-json==0.6.7
debugpy==1.8.8
decorator==5.1.1
deepdiff==8.0.1
distro==1.9.0
emoji==2.14.0
et_xmlfile==2.0.0
executing==2.1.0
filetype==1.2.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-openai==0.2.6
langchain-text-splitters==0.3.2
langchain-unstructured==0.1.5
langdetect==1.0.9
langgraph==0.2.45
langgraph-checkpoint==2.0.2
langgraph-sdk==0.1.35
langsmith==0.1.142
lxml==5.3.0
marshmallow==3.23.1
matplotlib-inline==0.1.7
msal==1.31.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
Office365-REST-Python-Client==2.5.13
openai==1.54.3
openpyxl==3.1.5
orderly-set==5.2.2
orjson==3.10.11
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pypdf==5.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.8.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.11.8
unstructured-client==0.25.9
urllib3==2.2.3
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.17.1
